Core/Misc: implement custom player stat boni
if you want to add an specific amout of any stat to a player you can use the player_bonus_stats table in zyn database.

just set the race and class IDs, specify the level and the boni amount for each stat.

this system is perfect if you don't feel like creating custom items for leveling purposes only

+++ src/server/game/Entities/Player/Player.h
@@ -229,9 +229,9 @@ struct PlayerClassInfo
 
 struct PlayerLevelInfo
 {
-    PlayerLevelInfo() { for (uint8 i=0; i < MAX_STATS; ++i) stats[i] = 0; }
+    PlayerLevelInfo() { for (uint16 i=0; i < MAX_STATS; ++i) stats[i] = 0; }
 
-    uint8 stats[MAX_STATS];
+    uint16 stats[MAX_STATS];
 };
 
 typedef std::list<uint32> PlayerCreateInfoSpells;
